+++ src/modules/tools/layerSlider/store/gettersLayerSlider.js
@@ -1,0 +1,9 @@
+
+import {generateSimpleGetters} from "../../../../app-store/utils/generators";
+import coordState from "./stateLayerSlider";
+
+const getters = {
+    ...generateSimpleGetters(coordState)
+};
+
+export default getters;