+++ src/utils/getWKTGeom.js
@@ -9,15 +9,35 @@ import {WKT} from "ol/format.js";
 function getWKTGeom (content, geometryType = "POLYGON") {
     const format = new WKT(),
         type = content?.geometryType ? content.geometryType : geometryType, // the default value is POLYGON because for type street, there is no geometryType defined. But it should be polygon
-        geometry = content?.coordinate ? content.coordinate : content,
-        interiorGeometry = content?.interiorGeometry;
+        geometry = content?.coordinate ? content.coordinate : content;
     let wkt,
         regExp;
 
     if (type === "POLYGON") {
         wkt = type + "((";
         geometry.forEach(function (element, index, list) {
-            if (index % 2 === 0) {
+            // polygon with voids
+            if (Array.isArray(element)) {
+                element.forEach(function (coord, index2, list2) {
+                    if (index2 % 2 === 0) {
+                        wkt += coord + " ";
+                    }
+                    else if (index2 === list2.length - 1) {
+                        wkt += coord + ")";
+                    }
+                    else {
+                        wkt += coord + ", ";
+                    }
+                });
+                if (index === list.length - 1) {
+                    wkt += ")";
+                }
+                else {
+                    wkt += ",(";
+                }
+            }
+            // simple polygon
+            else if (index % 2 === 0) {
                 wkt += element + " ";
             }
             else if (index === list.length - 1) {
@@ -35,26 +55,42 @@ function getWKTGeom (content, geometryType = "POLYGON") {
     }
     else if (type === "MULTIPOLYGON") {
         wkt = type + "(((";
+        // all single polygons
         geometry.forEach(function (element, index) {
-            geometry[index].forEach(function (coord, index2, list) {
-                if (index2 % 2 === 0) {
+            // goes through either all coordinates of a polygon or a polygon with voids
+            element.forEach(function (coord, index2, list2) {
+                // element is a polygon with voids
+                if (Array.isArray(coord)) {
+                    coord.forEach(function (coordinate, index3, list3) {
+                        if (index3 % 2 === 0) {
+                            wkt += coordinate + " ";
+                        }
+                        else if (index3 === list3.length - 1) {
+                            wkt += coordinate + ")";
+                        }
+                        else {
+                            wkt += coordinate + ", ";
+                        }
+                    });
+                    if (index2 === list2.length - 1) {
+                        wkt += ")";
+                    }
+                    else {
+                        wkt += ",(";
+                    }
+                }
+                // element is a simple polygon
+                else if (index2 % 2 === 0) {
                     wkt += coord + " ";
                 }
-                else if (index2 === list.length - 1 && interiorGeometry?.indexOf(index + 1) !== -1) {
-                    wkt += coord + ")";
-                }
-                else if (index2 === list.length - 1) {
+                else if (index2 === list2.length - 1) {
                     wkt += coord + "))";
                 }
                 else {
                     wkt += coord + ", ";
                 }
             });
-
-            if (interiorGeometry?.indexOf(index + 1) !== -1) {
-                wkt += ",(";
-            }
-            else if (index === geometry.length - 1 && interiorGeometry?.indexOf(index + 1) === -1) {
+            if (index === geometry.length - 1) {
                 wkt += ")";
             }
             else {